feature/YH-73:[Refactor] project structure fix
