+++ src/entities/question/api/__mocks__/getQuestionsListHandler.mock.ts
@@ -1,0 +1,7 @@
+import { http, HttpResponse } from 'msw';
+
+import { mockGetQuestionsList } from './mockGetQuestionsList';
+
+export const getQuestionsList = [
+	http.get('/questions', () => HttpResponse.json(mockGetQuestionsList)),
+];
