+++ src/entities/question/index.ts
@@ -1,3 +1,5 @@
 export type { Question } from './model/types/question';
 export { QuestionCard } from './ui/QuestionCard/QuestionCard';
 export { useGetQuestionByIdQuery, useGetQuestionsListQuery } from './api/questionsApi';
+export { getQuestionsList } from './api/__mocks__/getQuestionsListHandler.mock';
+export { getQuestionById } from './api/__mocks__/getQuestionByIdHandler.mock';
