--- src/entities/userSearch/index.ts
@@ -1,1 +1,0 @@
-export { useSearchUsersMutation } from './api/userSearchApi';
