+++ src/features/userAvatar/index.ts
@@ -1,0 +1,1 @@
+export { getUserAvatarById } from './api/__mocks__/getUserAvatarByIdHandler.mock';
