+++ src/features/userSearch/index.ts
@@ -1,0 +1,2 @@
+export { useCheckUsersExistsMutation } from './api/userSearchApi';
+export { checkUsersExists } from './api/__mocks__/checkUsersExistsHandler.mock';
